% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{benchmark_all}
\alias{benchmark_all}
\title{Benchmark all partition functions}
\usage{
benchmark_all(n = c(10, 100, 1000), times = 100L)
}
\arguments{
\item{n}{a vector of sample sizes to simulate for benchmarking. Default is to
test 10, 100, and 1000.}

\item{times}{number of benchmarks to run in \code{microbenchmark}. Default is 100.}
}
\value{
a \code{data.frame}
}
\description{
Benchmark \code{partition()} and \code{kmeans_icc()} using multiple sample sizes and
methods
}
