% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{benchmark_and_save}
\alias{benchmark_and_save}
\title{Benchmark and save data}
\usage{
benchmark_and_save(...)
}
\arguments{
\item{...}{arguments to \code{\link[=benchmark_all]{benchmark_all()}}}
}
\value{
invisibly returns a \code{data.frame}
}
\description{
Benchmark \code{partition()} and \code{kmeans_icc()} and save the results to the
internal \code{benchmarks} data.frame.
}
