% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{benchmark_partition}
\alias{benchmark_partition}
\alias{benchmark_kmeans}
\title{Benchmark specific partition functions}
\usage{
benchmark_partition(n = c(10, 100, 1000), pct.var = 0.8, method = "ICC",
  dist.type = "p", unit = "ms", times = 100L)

benchmark_kmeans(n = c(10, 100, 1000), threshold.icc = 0.4, ...,
  unit = "ms", times = 100L)
}
\arguments{
\item{n}{a vector of sample sizes to simulate for benchmarking. Default is to
test 10, 100, and 1000.}

\item{pct.var}{see \link[partition:partition]{partition::partition}}

\item{method}{see \link[partition:partition]{partition::partition}}

\item{dist.type}{see \link[partition:partition]{partition::partition}}

\item{unit}{what unit of time? Default is "ms" for milliseconds. See
\link[microbenchmark:microbenchmark]{microbenchmark::microbenchmark}}

\item{times}{number of benchmarks to run in \code{microbenchmark}. Default is 100.}

\item{threshold.icc}{see \link[partition:kmeans_icc]{partition::kmeans_icc}}
}
\value{
a \code{data.frame}
}
\description{
Benchmark \code{partition()} and \code{kmeans_icc()} using multiple sample sizes and
methods
}
