% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{plot_benchmarks}
\alias{plot_benchmarks}
\title{Plot stored benchmarks}
\usage{
plot_benchmarks(what = "all", .df = NULL)
}
\arguments{
\item{what}{What should be plotted? Either "kmeans_icc()" or "partition()"}

\item{.df}{a \code{data.frame} with benchmarks. Default is \code{NULL}, which plots
benchmarks in the internal \code{benchmarks} data.}
}
\value{
a \code{ggplot}
}
\description{
\code{plot_benchmarks()} uses \code{ggplot2} and \code{ggridges} to plot densities of
benchmarked times in the stored \code{benchmarks} data or new data.
}
